pyproject.toml
[project]
name = "mcp-go-test"
version = "0.1.0"
description = "Test scripts for mcp-go server"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.14",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
